pyproject.toml
[build-system] 
requires = ["setuptools" ] 
build-backend = "setuptools.build_meta" 

[project] 
name = "noise-barrier-intel" 
version = "0.1.0"
description = "A monorepo for crawling and agent-based data extraction."
dependencies = [
    "requests",
    "beautifulsoup4",
    "SQLAlchemy",
    "pymysql",
    "python-dotenv",
    "langgraph",
    "langchain-anthropic",
    "langchain-mcp-adapters",
    "playwright",
    "apscheduler",
    "pandas"
]

[tool.setuptools.packages.find] 
where = ["."]